Fix modprobe location resolution
Ansible's `first_found` lookup plugin runs in the controller host, not in the remote hosts.
Therefore, if the `modprobe` location in the target host is not the same as on the remote host,
the role fails silently since it does not find the `modprobe` binary and creates a syntax error
in the systemd unit file.

This approach uses `stat` instead to probe for the `modprobe` binary in various locations and then
selects the first found. This has the added benefit of failing when the binary is not found in the
remote host since the `first` filter fails when the list is empty.